Create README.md
Ive build this site but I am having trouble getting it to scale so that it is viewable on various devices. I would like to use media queries to do this. Ive attempted to use media queries but they don't seem work. I may have written the html and css in a way that makes this difficult.